intel_service/requirements.txt
# Phase 7: scrapers + scheduler + DB writes
requests>=2.28.0
beautifulsoup4>=4.11.0
APScheduler>=3.10.0
loguru>=0.6.0
PyYAML>=6.0
